+++ backends/vulkan/runtime/api/Descriptor.h
@@ -14,8 +14,8 @@
 
 #include <executorch/backends/vulkan/runtime/api/Shader.h>
 
-#include <executorch/backends/vulkan/runtime/api/resource/Buffer.h>
-#include <executorch/backends/vulkan/runtime/api/resource/Image.h>
+#include <executorch/backends/vulkan/runtime/api/memory/Buffer.h>
+#include <executorch/backends/vulkan/runtime/api/memory/Image.h>
 
 #include <unordered_map>
