+++ backends/vulkan/runtime/api/api.h
@@ -12,6 +12,7 @@
 #include <executorch/backends/vulkan/runtime/api/Command.h>
 #include <executorch/backends/vulkan/runtime/api/Context.h>
 #include <executorch/backends/vulkan/runtime/api/Descriptor.h>
+#include <executorch/backends/vulkan/runtime/api/Fence.h>
 #include <executorch/backends/vulkan/runtime/api/Pipeline.h>
 #include <executorch/backends/vulkan/runtime/api/Runtime.h>
 #include <executorch/backends/vulkan/runtime/api/Shader.h>
@@ -19,8 +20,7 @@
 #include <executorch/backends/vulkan/runtime/api/Tensor.h>
 #include <executorch/backends/vulkan/runtime/api/Utils.h>
 
-#include <executorch/backends/vulkan/runtime/api/resource/Allocation.h>
-#include <executorch/backends/vulkan/runtime/api/resource/Allocator.h>
-#include <executorch/backends/vulkan/runtime/api/resource/Buffer.h>
-#include <executorch/backends/vulkan/runtime/api/resource/Fence.h>
-#include <executorch/backends/vulkan/runtime/api/resource/Image.h>
+#include <executorch/backends/vulkan/runtime/api/memory/Allocation.h>
+#include <executorch/backends/vulkan/runtime/api/memory/Allocator.h>
+#include <executorch/backends/vulkan/runtime/api/memory/Buffer.h>
+#include <executorch/backends/vulkan/runtime/api/memory/Image.h>
