+++ backends/vulkan/runtime/api/memory/Allocation.h
@@ -12,15 +12,11 @@
 
 #include <executorch/backends/vulkan/runtime/api/vk_api.h>
 
-#include <executorch/backends/vulkan/runtime/api/Types.h>
-#include <executorch/backends/vulkan/runtime/api/Utils.h>
+#include <executorch/backends/vulkan/runtime/api/Exception.h>
 
-#include <executorch/backends/vulkan/runtime/api/resource/vma_api.h>
+#include <executorch/backends/vulkan/runtime/api/memory/vma_api.h>
 
-#include <mutex>
 #include <ostream>
-#include <stack>
-#include <unordered_map>
 
 std::ostream& operator<<(std::ostream& out, VmaTotalStatistics stats);
