+++ backends/vulkan/runtime/api/memory/Allocator.cpp
@@ -6,9 +6,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-#include <executorch/backends/vulkan/runtime/api/Adapter.h>
-
-#include <executorch/backends/vulkan/runtime/api/resource/Allocator.h>
+#include <executorch/backends/vulkan/runtime/api/memory/Allocator.h>
 
 namespace vkcompute {
 namespace api {
