+++ backends/vulkan/runtime/api/memory/Buffer.h
@@ -12,17 +12,11 @@
 
 #include <executorch/backends/vulkan/runtime/api/vk_api.h>
 
-#include <executorch/backends/vulkan/runtime/api/Types.h>
 #include <executorch/backends/vulkan/runtime/api/Utils.h>
 
-#include <executorch/backends/vulkan/runtime/api/resource/vma_api.h>
+#include <executorch/backends/vulkan/runtime/api/memory/vma_api.h>
 
-#include <executorch/backends/vulkan/runtime/api/resource/Allocation.h>
-
-#include <mutex>
-#include <ostream>
-#include <stack>
-#include <unordered_map>
+#include <executorch/backends/vulkan/runtime/api/memory/Allocation.h>
 
 namespace vkcompute {
 namespace api {
